general work

+++ src/graphics/graphics.c
@@ -5,5 +5,5 @@
 
 unsigned short g_xres;
 unsigned short g_yres;
-double g_aspect_ratio;
+float g_aspect_ratio;
 bool g_fullscreen = false;
